refactor: move app metadata out of paths
