+++ src/Fleasion/utils/metadata.py
@@ -1,0 +1,12 @@
+"""Application metadata."""
+
+from Fleasion import __version__ as APP_VERSION
+
+
+APP_NAME = 'Fleasion'
+APP_AUTHOR = '@8ar__, @dis_spencer, @1_v'
+APP_LOGIC = '@blockce, @0100152000022000, @Yeha., @emk530'
+APP_CONCEPT = '@cro.p'
+APP_DISCORD = 'discord.gg/hXyhKehEZF'
+APP_REPO = 'https://github.com/fleasion/Fleasion'
+
