+++ src/Fleasion/utils/paths.py
@@ -1,18 +1,10 @@
 """Application paths and constants."""
 
-import sys
 import os
+import sys
 from pathlib import Path
 
-from Fleasion import __version__ as APP_VERSION
-
-# Application metadata
-APP_NAME = 'Fleasion'
-APP_AUTHOR = '@8ar__, @dis_spencer, @1_v'
-APP_LOGIC = '@blockce, @0100152000022000, @Yeha., @emk530'
-APP_CONCEPT = '@cro.p'
-APP_DISCORD = 'discord.gg/hXyhKehEZF'
-APP_REPO = 'https://github.com/fleasion/Fleasion'
+from Fleasion.utils.metadata import APP_NAME
 
 # Process and proxy configuration
 if sys.platform == 'darwin':
